Centralizes release pagination types in shared module
Unifies release pagination and query parameter types across backend and frontend
by moving their definitions to the shared module. Reduces duplication, ensures
type consistency, and simplifies maintenance when updating release-related API
contracts.
